修改数据库重新生成类

+++ springboot_demos/demo_interceptor/src/main/java/top/actualself/springboot/demo_interceptor/service/UserService.java
@@ -6,11 +6,11 @@ import top.actualself.springboot.demo_interceptor.model.User;
 import java.util.List;
 
 public interface UserService {
-    User getUserById(Long id);
+    User getUserById(Integer id);
 
     List<User> getAllUser(Integer page, Integer row);
 
     WidelyResult modifyUser(User user);
 
-    WidelyResult delUserById(Long id);
+    WidelyResult delUserById(Integer id);
  }
